requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-dotenv>=1.0.0
httpx>=0.24.0
numpy>=1.24.0
python-telegram-bot>=21.0
reportlab>=4.0.0